Huge performance improvements in batch preparations.
All images are now stored in pinned memory to be loaded much faster. This comes at the cost of a factor of 2 less available RAM for the replay buffer. Cuda streams are used to allow the overlap of batch preparation and learning/inference. A modified version of torchrl's ReplayBuffer is used to synchronize() minimally.